🛠️refactoring typing to interface
