+++ src/commands/conversation/delete.ts
@@ -1,6 +1,6 @@
 import { window } from 'vscode'
 import { Command } from '../commandManager'
-import { IConversation } from '@app/types'
+import { IConversation } from '@app/interfaces'
 import { ConversationStorageService } from '@app/services'
 
 export default class DeleteConversationCommand implements Command {
