+++ src/commands/conversation/openMarkdown.ts
@@ -1,6 +1,6 @@
 import { ViewColumn, window, workspace } from 'vscode'
 import { Command } from '../commandManager'
-import { IConversation } from '@app/types'
+import { IConversation } from '@app/interfaces'
 
 export default class ShowConversationMarkdownCommand implements Command {
   public readonly id = '_vscode-openai.conversation.show.markdown'
