+++ src/commands/conversation/openWebview.ts
@@ -1,5 +1,5 @@
 import { Command } from '../commandManager'
-import { IConversation } from '@app/types'
+import { IConversation } from '@app/interfaces'
 import { ConversationStorageService } from '@app/services'
 
 export default class OpenConversationWebviewCommand implements Command {
