+++ src/commands/embeddings/newConversationEmbeddingAll.ts
@@ -1,5 +1,5 @@
 import { Command } from '../commandManager'
-import { IConversation } from '@app/types'
+import { IConversation } from '@app/interfaces'
 import { ConversationStorageService } from '@app/services'
 import { getQueryResourcePersona } from '@app/models'
 import { VSCODE_OPENAI_EMBEDDING } from '@app/constants'
