+++ src/commands/messages/copyClipboardMessage.ts
@@ -1,7 +1,7 @@
 import { env } from 'vscode'
 import { Command } from '../commandManager'
 import { showMessageWithTimeout } from '@app/apis/vscode'
-import { IChatCompletion } from '@app/types'
+import { IChatCompletion } from '@app/interfaces'
 
 export default class ClipboardCopyMessagesMessageCommand implements Command {
   public readonly id = '_vscode-openai.messages.clipboard-copy.message'
