+++ src/interfaces/ICodeDocument.ts
