+++ src/interfaces/IConfigurationConversation.ts
