+++ src/interfaces/IConfigurationConversationColor.ts
