+++ src/interfaces/IConfigurationEmbedding.ts
