+++ src/interfaces/IConfigurationSetting.ts
