+++ src/interfaces/IConversation.ts
@@ -1,4 +1,4 @@
-import { IPersonaOpenAI, IChatCompletion } from '@app/types'
+import { IPersonaOpenAI, IChatCompletion } from '@app/interfaces'
 
 export interface IConversation {
   timestamp: number
