+++ src/interfaces/IDynamicLooseObject.ts
