+++ src/interfaces/IEmbeddingFileChunk.ts
