+++ src/interfaces/IEmbeddingFileLite.ts
