+++ src/interfaces/IEmbeddingText.ts
