+++ src/interfaces/IMessage.ts
