+++ src/interfaces/IPersonaOpenAI.ts
