+++ src/utilities/editor/compareResultsToClipboard.ts
@@ -1,6 +1,6 @@
 import { commands, env, window } from 'vscode'
 import { ConversationStorageService } from '@app/services'
-import { IChatCompletion, IConversation, IPersonaOpenAI } from '@app/types'
+import { IChatCompletion, IConversation, IPersonaOpenAI } from '@app/interfaces'
 import { createChatCompletion } from '@app/apis/openai'
 import { ChatCompletionConfigFactory } from '@app/services/configurationServices'
